+++ mail/mail-common/src/draft.rs
@@ -23,7 +23,7 @@ use proton_action_queue::action::{ActionId, MetadataBuilder};
 use proton_action_queue::queue::{ActionError, Queue, QueuedActionOutput};
 use proton_api_core::consts::Mail;
 use proton_api_core::service::ApiServiceError;
-use proton_api_core::services::proton::common::{AddressId, AuthId};
+use proton_api_core::services::proton::common::{AddressId, SessionId};
 use proton_api_core::session::{CoreSession, Session};
 use proton_api_mail::services::proton::prelude::DraftReplyOrForwardParams;
 use proton_api_mail::services::proton::request_data::{DraftAction, DraftAttachmentKeyPackets};
@@ -654,7 +654,7 @@ impl Draft {
         source_message: &Message,
         source_message_body: DecryptedMessageBody,
         use_utc: bool,
-        session_id: &AuthId,
+        session_id: &SessionId,
     ) -> (Self, Vec<Attachment>) {
         let mut body = get_signature(address, mail_settings);