+++ mail/mail-uniffi/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased] - 2025-03-xx
 
+### Changed
+  - Rename `auth_id` to `session_id` in `EncryptedPushNotification` type.
+
+
 ## [0.56.0] - 2025-03-11
 
 ### Added